Create method disemvowel
Created method disemvowel that takes the arguments of an inputFile and an outputFile. Then it uses a while loop to read from the inputFile into the in_buffer. Then call the method copy_non_vowels and then finally write the output from the copy_non_vowels method to the outputFile.